Don't assume sizeof(void *) == sizeof(unsigned long).

+++ etc/afpd/desktop.c
@@ -856,7 +856,7 @@ int afp_addcomment(AFPObj *obj, char *ibuf, size_t ibuflen _U_, char *rbuf _U_, 
 	return get_afp_errno(AFPERR_NOOBJ);
     }
 
-    if ((u_long)ibuf & 1 ) {
+    if ((intptr_t)ibuf & 1 ) {
         ibuf++;
     }
